+++ controllers/preview_controller/previewers/http.go
@@ -20,7 +20,7 @@ import (
 	"github.com/turt2live/matrix-media-repo/controllers/preview_controller/preview_types"
 )
 
-func doHttpGet(urlPayload *preview_types.UrlPayload, ctx rcontext.RequestContext) (*http.Response, error) {
+func doHttpGet(urlPayload *preview_types.UrlPayload, languageHeader string, ctx rcontext.RequestContext) (*http.Response, error) {
 	var client *http.Client
 
 	dialer := &net.Dialer{
@@ -116,12 +116,13 @@ func doHttpGet(urlPayload *preview_types.UrlPayload, ctx rcontext.RequestContext
 		return nil, err
 	}
 	req.Header.Set("User-Agent", "matrix-media-repo")
+	req.Header.Set("Accept-Language", languageHeader)
 	return client.Do(req)
 }
 
-func downloadRawContent(urlPayload *preview_types.UrlPayload, supportedTypes []string, ctx rcontext.RequestContext) ([]byte, string, string, string, error) {
+func downloadRawContent(urlPayload *preview_types.UrlPayload, supportedTypes []string, languageHeader string, ctx rcontext.RequestContext) ([]byte, string, string, string, error) {
 	ctx.Log.Info("Fetching remote content...")
-	resp, err := doHttpGet(urlPayload, ctx)
+	resp, err := doHttpGet(urlPayload, languageHeader, ctx)
 	if err != nil {
 		return nil, "", "", "", err
 	}
@@ -164,8 +165,8 @@ func downloadRawContent(urlPayload *preview_types.UrlPayload, supportedTypes []s
 	return bytes, filename, contentType, resp.Header.Get("Content-Length"), nil
 }
 
-func downloadHtmlContent(urlPayload *preview_types.UrlPayload, supportedTypes []string, ctx rcontext.RequestContext) (string, error) {
-	raw, _, _, _, err := downloadRawContent(urlPayload, supportedTypes, ctx)
+func downloadHtmlContent(urlPayload *preview_types.UrlPayload, supportedTypes []string, languageHeader string, ctx rcontext.RequestContext) (string, error) {
+	raw, _, _, _, err := downloadRawContent(urlPayload, supportedTypes, languageHeader, ctx)
 	html := ""
 	if raw != nil {
 		html = string(raw)
@@ -173,9 +174,9 @@ func downloadHtmlContent(urlPayload *preview_types.UrlPayload, supportedTypes []
 	return html, err
 }
 
-func downloadImage(urlPayload *preview_types.UrlPayload, ctx rcontext.RequestContext) (*preview_types.PreviewImage, error) {
+func downloadImage(urlPayload *preview_types.UrlPayload, languageHeader string, ctx rcontext.RequestContext) (*preview_types.PreviewImage, error) {
 	ctx.Log.Info("Getting image from " + urlPayload.ParsedUrl.String())
-	resp, err := doHttpGet(urlPayload, ctx)
+	resp, err := doHttpGet(urlPayload, languageHeader, ctx)
 	if err != nil {
 		return nil, err
 	}