+++ migrations/15_add_language_url_previews_down.sql
@@ -1,0 +1,1 @@
+ALTER TABLE url_previews DROP COLUMN language_header;