+++ migrations/15_add_language_url_previews_up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE url_previews ADD COLUMN language_header TEXT NULL DEFAULT NULL;