+++ types/url_preview.go
@@ -1,16 +1,17 @@
 package types
 
 type UrlPreview struct {
-	Url         string
-	SiteName    string
-	Type        string
-	Description string
-	Title       string
-	ImageMxc    string
-	ImageType   string
-	ImageSize   int64
-	ImageWidth  int
-	ImageHeight int
+	Url            string
+	SiteName       string
+	Type           string
+	Description    string
+	Title          string
+	ImageMxc       string
+	ImageType      string
+	ImageSize      int64
+	ImageWidth     int
+	ImageHeight    int
+	LanguageHeader string
 }
 
 type CachedUrlPreview struct {