+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 * Added support for [MSC2448](https://github.com/matrix-org/matrix-doc/pull/2448).
 * Added support for a `forKinds: ["all"]` option on datastores.
 * Added support for specifying a `region` to the S3 provider.
+* Pass-through the `Accept-Language` header for URL previews, with options to set a default.
 
 ### Fixed